feature/CA-002 refactorizacion del codigo en componentes
