+++ decide/booth/decide-react/src/Barra.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import Navbar from 'react-bootstrap/Navbar'
 import Nav from 'react-bootstrap/Nav'
